% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{age_imp_preprocessing}
\alias{age_imp_preprocessing}
\title{Impute age values of trees}
\usage{
age_imp_preprocessing(data)
}
\arguments{
\item{data}{\link{data.frame} containing a variable named "age"}
}
\description{
Impute age values of trees
}
