% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{age_imputation}
\alias{age_imputation}
\title{Imputation of age values}
\usage{
age_imputation(x, y, age, print = TRUE)
}
\arguments{
\item{x}{X coordinate}

\item{y}{Y coordinate}

\item{age}{Age value}
}
\description{
Imputation of age values
}
\details{
This function imputes age values of trees assuming that a tree is
very likely to have the same age as the tree it is surrounded by.
}
