% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{atlas_climatico_download}
\alias{atlas_climatico_download}
\title{Download Atlas Climatico data}
\usage{
atlas_climatico_download(url)
}
\arguments{
\item{url}{URL to download from (usually Zenodo)}
}
\description{
Download Atlas Climatico data
}
