% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{atlas_climatico_preprocessing}
\alias{atlas_climatico_preprocessing}
\title{## Atlas Climatico}
\usage{
atlas_climatico_preprocessing(path, study_area)
}
\description{
Raw data: Zipped raster files of temperature, pisr and precipitation in `.txt` format in CRS 23030.
}
\details{
Workflow: Unzip files specified by pattern.
          Remove unused `.gif` and `.htm` files.
          Reproject study area to CRS 23030 to crop and mask raster files.
          Reproject raster files to CRS 32630.
          Write raster files to disk.
          Remove unused `.txt` files

Mod data: Temperature, pisr and precipitation rasters in `.tif` format in CRS 32630 for each month.
}
