% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-functions.R
\name{benchmark_custom}
\alias{benchmark_custom}
\title{mlr::benchmark() wrapper}
\usage{
benchmark_custom(task, learner, resampling)
}
\arguments{
\item{task}{mlr task}

\item{learner}{Learner from \link[mlr:makeLearner]{mlr::makeLearner}}

\item{resampling}{mlr resampling}
}
\description{
mlr::benchmark() wrapper
}
