% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_dataset_tables}
\alias{create_dataset_tables}
\title{Create continuos and nominal LaTeX tables from dataset}
\usage{
create_dataset_tables(data, type = "continuous", drop_vars = NULL,
  stats = NULL)
}
\arguments{
\item{data}{\link{data.frame}}

\item{type}{Options: "continuous" or "nominal"}
}
\description{
Create continuos and nominal LaTeX tables from dataset
}
