% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{elevation_preprocessing}
\alias{elevation_preprocessing}
\title{Raw data: Zipped 5m lidar DEM in CRS 3042 in `.tif` format.}
\usage{
elevation_preprocessing(path)
}
\description{
Workflow: Reprojection on CRS 32630. Extraction of the elevation values for each observation.
}
\details{
Mod data: DEM with 5m resolution in `.tif` format in CRS 32630.
}
