% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{extract_variables}
\alias{extract_variables}
\title{Extract variables to in-situ data}
\usage{
extract_variables(data, temperature, precipitation, pisr, slope, elevation,
  soil, lithology, hail, ph, study_area, response, drop_vars = NULL,
  age = FALSE)
}
\arguments{
\item{data}{Input \link{data.frame} containing in-situ points. Here a Geopackage
file downloaded directly from Zenodo.}

\item{temperature}{Temperature data (mean)}

\item{precipitation}{Precipitation data (sum)}

\item{pisr}{Potential incoming solar radiation (as a fraction of its mean)}

\item{slope}{Slope (in degrees)}

\item{elevation}{Elevation in meters}

\item{soil}{Soil type}

\item{lithology}{Lithology type}

\item{hail}{Probability of hail damage at trees}

\item{ph}{pH value of soil}

\item{study_area}{study area}

\item{response}{Name of the response variable}

\item{drop_vars}{A vector of variables to drop}

\item{age}{(logical) Whether "age imputation" should be performed}
}
\description{
Extract variables to in-situ data
}
\details{
This function extract all variables to the data points.
Variables "temperature", "precipitation" and "pisr" already existed in the
dataset and were not extracted to the points.
}
\seealso{
\link{extract_variables_v2}
}
