% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{lithology_download}
\alias{lithology_download}
\title{Download lithology data}
\usage{
lithology_download(url)
}
\arguments{
\item{url}{URL to download from (usually Zenodo)}
}
\description{
Download lithology data
}
