% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{lithology_preprocessing}
\alias{lithology_preprocessing}
\title{Raw data: Polygon shapefile with lithology units in CRS 25830.}
\usage{
lithology_preprocessing(lithology)
}
\description{
Workflow: Read shapefile and reproject to CRS 32630.
          Extract column with lithology unit id.
          Make lithology sf object valid and extract at each observation the lithology unit id.
          Aggregate lithology units to common lithology classes.
}
\details{
Output: Aggregated lithology units.
}
