% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{mod_raw_data}
\alias{mod_raw_data}
\title{Modification of raw data}
\usage{
mod_raw_data(data, drop_vars, response)
}
\arguments{
\item{data}{Raw input data}

\item{drop_vars}{Variable to remove from the data}

\item{response}{Variable that will be the response of the task later}
}
\description{
Modification of raw data
}
\details{
The following modifications are handled in this function:
\itemize{
\item Transformation of variable names to lowercase
\item Wiggling of coordinates so that they are "unique" in a spatial sense
\item Custom replacement of specific variables using expert knowledge
\item Removal of points with no coordinate information
\item Filtering of observations to a certain tree species
}
}
