% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{ph_preprocessing}
\alias{ph_preprocessing}
\title{Preprocessing of pH data}
\usage{
ph_preprocessing(data, study_area)
}
\arguments{
\item{data}{Downloaded pH data}
}
\description{
Preprocessing of pH data
}
