% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{pisr_preprocessing}
\alias{pisr_preprocessing}
\title{## PISR}
\usage{
pisr_preprocessing(atlas_climatico)
}
\description{
Raw data: PISR rasters for month july to september in `.tif`´format in CRS 32630.
}
\details{
Workflow: Get files by pattern and stack files.
          Add pisr of july to september and convert to the relative fraction of its mean (mean = 8065.952 kW/m^2).
          Write raster to disk and extract pisr at each observation.

Modified data: PISR raster in `.tif` format in CRS 32630.
}
