% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{precipitation_preprocessing}
\alias{precipitation_preprocessing}
\title{## Precipitation}
\usage{
precipitation_preprocessing(atlas_climatico)
}
\description{
Raw data: Precipitation rasters for month july and august in `.tif`´format in CRS 32630.
}
\details{
Workflow: Get files by pattern and stack files.
          Add precipitation of month july and august and write raster to disk.
          Extract precipitation at each observation.

Modified data: Total july and august precipitation raster in `.tif` format in CRS 32630.
}
