% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-functions.R
\name{prediction_custom}
\alias{prediction_custom}
\title{Predict infection state of pathogens}
\usage{
prediction_custom(..., learner, param_set, tune_ctrl, resampling,
  prediction_data, prediction_grid, desc_resampling)
}
\arguments{
\item{...}{mlr task objects}

\item{learner}{Learner from \link[mlr:makeLearner]{mlr::makeLearner}}

\item{param_set}{mlr Param Set}

\item{tune_ctrl}{mrl tune control object}

\item{resampling}{mlr resampling}

\item{prediction_data}{Prediction data.frame}

\item{prediction_grid}{Raster Grid to which the prediction should be mapped to}

\item{desc_resampling}{Resampling description}
}
\description{
Predict infection state of pathogens
}
