% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{preprocessing_custom}
\alias{preprocessing_custom}
\title{# Remove samples with NAs
Examine datasets for missing values. Remove sample 12 from heterobasi data set because lithology is missing
and remove sample 11, 13, 14, 15, 22, 23, 25, 26, 27 and 28 from armillaria data set because elevation and/or lithology is missing.}
\usage{
preprocessing_custom(path, slope, soil, temperature_mean, ph, hail,
  precipitation_sum, elevation, pisr, lithology,
  study_area = data_basque, response, drop_vars = NULL, age = FALSE)
}
\description{
# Remove samples with NAs
Examine datasets for missing values. Remove sample 12 from heterobasi data set because lithology is missing
and remove sample 11, 13, 14, 15, 22, 23, 25, 26, 27 and 28 from armillaria data set because elevation and/or lithology is missing.
}
