% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{remove_extract}
\alias{remove_extract}
\title{remove_extract}
\usage{
remove_extract(bm_result)
}
\arguments{
\item{bm_result}{BenchmarkResult object}
}
\description{
Removes the extract column of a \code{mlr} BenchmarkResult object
}
\details{
Typically, tuning results are stored in this slot. Combining many benchmark
results can lead to a huge object. For visualizing the performance, the
object can be shrinked with this function.
}
