% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{save_plot_custom}
\alias{save_plot_custom}
\title{Save plot and convert to pdf}
\usage{
save_plot_custom(plot, path, nrow, ncol)
}
\arguments{
\item{plot}{A ggplot2 plot}

\item{path}{The path where to save the plot (including filename)}

\item{nrow}{number of rows}

\item{ncol}{number of columns}
}
\description{
Save plot and convert to pdf
}
\details{
Same behaviour as \link[cowplot:save_plot]{cowplot::save_plot}
}
