% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{save_plot_opt_paths}
\alias{save_plot_opt_paths}
\title{Write optimization path figures to disk}
\usage{
save_plot_opt_paths(list)
}
\arguments{
\item{list}{A list generated from \code{vis_opt_path()}}
}
\description{
Write optimization path figures to disk
}
