% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{slope_processing}
\alias{slope_processing}
\title{## Slope}
\usage{
slope_processing(path)
}
\description{
Raw data: Reprojected DEM with 5m resolution in `.tif` format in CRS 32630.
}
\details{
Workflow: Calculation of the slope with the RSAGA package.
          Conversion of the resulting `.sgrd` to `.tif`.
          Extraction of the slope values for each record.

Mod data: Slope raster in degrees with 5m resolution in `.tif` format in CRS 32630.
}
