% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{soil_preprocessing}
\alias{soil_preprocessing}
\title{## Soil}
\usage{
soil_preprocessing(path, study_area)
}
\description{
Raw data: Soil raster with 250m resolution in `.tif`format in CRS 4326 (Hengl 2017, http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0169748).
}
\details{
Workflow: Reproject study area to CRS 4326.
          Crop soil raster with study area and reproject to CRS 23030.
          Extraction of the soil value for each record.
          Left join with soil database from GSIF package to get actual soil name.
          Remove unused columns and factor soil name.

Output: Soil raster with 250 m resolution in `.tif` format in CRS 23030.
}
