% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{task_custom_prediction}
\alias{task_custom_prediction}
\title{Create a \link[mlr:task]{mlr::task} for performance estimation}
\usage{
task_custom_prediction(data, name, target, dummy_features,
  dummy.factors = FALSE, remove.vars = FALSE)
}
\arguments{
\item{data}{\link{data.frame}}

\item{name}{ID of the task}

\item{target}{Response of the task}

\item{dummy_features}{(logical) Whether factor levels should be dummy encoded}

\item{dummy.factors}{Which factor levels should be dummy encoded}
}
\description{
Create a \link[mlr:task]{mlr::task} for performance estimation
}
\details{
Differs to \link{task_custom} in that it removes variables
that are not available for prediction
}
