% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{temperature_preprocessing}
\alias{temperature_preprocessing}
\title{## Temperature}
\usage{
temperature_preprocessing(atlas_climatico)
}
\description{
Raw data: Temperature rasters for month march to september in `.tif`´format in CRS 32630.
}
\details{
Workflow: Get files by pattern and stack files.
          Calculate mean temperature of month march to september and write raster to disk.
          Extract mean summer temperature at each observation.

Modified data: Mean summer temperature raster in `.tif` format in CRS 32630.
}
