% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr-functions.R
\name{tuning_wrapper_custom}
\alias{tuning_wrapper_custom}
\title{mlr tuning wrapper}
\usage{
tuning_wrapper_custom(lrn, param_set, resampling, tune_ctrl)
}
\arguments{
\item{param_set}{mlr Param Set}

\item{resampling}{mlr resampling}

\item{tune_ctrl}{mrl tune control object}

\item{learner}{Learner from \link[mlr:makeLearner]{mlr::makeLearner}}
}
\description{
mlr tuning wrapper
}
