% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vis_partitions}
\alias{vis_partitions}
\title{Visualization of partitions used in benchmarking}
\usage{
vis_partitions(data, resampling_sp, resampling_nsp)
}
\arguments{
\item{resampling_sp}{A benchmark/resampling object with a spatial resampling}

\item{resampling_nsp}{A benchmark/resampling object with a non-spatial resampling}
}
\description{
Visualization of partitions used in benchmarking
}
