% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vis_tuning_effects}
\alias{vis_tuning_effects}
\title{Visualization of tuning effects}
\usage{
vis_tuning_effects(models, model_name, resampling, hyperparameter, xlim,
  ylim, default)
}
\arguments{
\item{models}{List of Benchmark/Resample results}

\item{model_name}{Name of algorithm for title}

\item{resampling}{String specifying the resample setting}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{default}{Default hyperparameter settings (x and y). A red cross will
denote the default settings in the plot.}

\item{hyperparam}{String of 2L specyfying the hyperparamters to compare}
}
\description{
Visualization of tuning effects
}
