% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{atlas_climatico_preprocessing}
\alias{atlas_climatico_preprocessing}
\title{Preprocessing of Atlas Climatico data}
\usage{
atlas_climatico_preprocessing(data, study_area)
}
\arguments{
\item{data}{Downloaded Atlas Climatico data}
}
\description{
Preprocessing of Atlas Climatico data
}
