% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{create_prediction_data}
\alias{create_prediction_data}
\title{Make prediction data.frame}
\usage{
create_prediction_data(temperature, precipitation, pisr, slope, elevation,
  soil, lithology, hail, ph)
}
\description{
Make prediction data.frame
}
