% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{elevation_preprocessing}
\alias{elevation_preprocessing}
\title{Preprocessing of elevation data}
\usage{
elevation_preprocessing(data)
}
\arguments{
\item{data}{Downloaded elevation data}
}
\description{
Preprocessing of elevation data
}
