% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{lithology_preprocessing}
\alias{lithology_preprocessing}
\title{Preprocessing of lithology data}
\usage{
lithology_preprocessing(data)
}
\arguments{
\item{data}{Downloaded lithology data}
}
\description{
Preprocessing of lithology data
}
