% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{mod_raw_data}
\alias{mod_raw_data}
\title{## Import disease data and store as sf}
\usage{
mod_raw_data(data, drop_vars, response)
}
\description{
Raw data: Point shapefile in CRS 23030 with presense and absence observations of heterobasi and armillaria.
}
\details{
## Change column names to lowercase
## Wiggle coordinates
Some coordinates are not unique, so we have to wiggle them a little bit
}
