% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{pisr_preprocessing}
\alias{pisr_preprocessing}
\title{Preprocessing of PISR data}
\usage{
pisr_preprocessing(atlas_climatico)
}
\arguments{
\item{atlas_climatico}{Preprocessed Atlas Climatico data}
}
\description{
Preprocessing of PISR data
}
