% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{precipitation_preprocessing}
\alias{precipitation_preprocessing}
\title{Preprocessing of precipitation data}
\usage{
precipitation_preprocessing(atlas_climatico)
}
\arguments{
\item{atlas_climatico}{Preprocessed Atlas Climatico data}
}
\description{
Preprocessing of precipitation data
}
