% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-functions.R
\name{prediction_custom}
\alias{prediction_custom}
\title{# Visualize all predictors}
\usage{
prediction_custom(..., learner, param_set, tune_ctrl, resampling,
  prediction_data, prediction_grid, desc_resampling)
}
\description{
# Visualize all predictors
}
