% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{slope_processing}
\alias{slope_processing}
\title{Preprocessing of slope data}
\usage{
slope_processing(data)
}
\arguments{
\item{data}{Downloaded DEM data}
}
\description{
Preprocessing of slope data
}
