% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{soil_preprocessing}
\alias{soil_preprocessing}
\title{Preprocessing of soil data}
\usage{
soil_preprocessing(data, study_area)
}
\arguments{
\item{data}{Downloaded soil data}

\item{study_area}{Geopackage of study area}
}
\description{
Preprocessing of soil data
}
