% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{temperature_preprocessing}
\alias{temperature_preprocessing}
\title{Preprocessing of temperature data}
\usage{
temperature_preprocessing(atlas_climatico)
}
\arguments{
\item{atlas_climatico}{Preprocessed Atlas Climatico data}
}
\description{
Preprocessing of temperature data
}
