% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vis_opt_path}
\alias{vis_opt_path}
\title{Visualization of optimization paths}
\usage{
vis_opt_path(models)
}
\arguments{
\item{model2}{List of Benchmark/Resample results}

\item{n_folds}{Number of folds to visualize (e.g. 1-5)}
}
\description{
Visualization of optimization paths
}
