% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vis_tuning_effects}
\alias{vis_tuning_effects}
\title{Visualization of tuning effects}
\usage{
vis_tuning_effects(models, resampling, model_name, hyperparam)
}
\arguments{
\item{models}{List of Benchmark/Resample results}

\item{resampling}{String specifying the resample setting}

\item{model_name}{Name of algorithm for title}

\item{hyperparam}{String of 2L specyfying the hyperparamters to compare}
}
\description{
Visualization of tuning effects
}
